Refactor tests by moving them to a separate file for better organization and maintainability
